Refatora fetch de licitações
